Added timestamp tag (element) for version control and changed all table IDs to "Long"
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@799 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/dbsupport/AttributeIFace.java
@@ -60,13 +60,13 @@ public interface AttributeIFace
      * Returns the record id.
      * @return the record id
      */
-    public Integer getAttrId();
+    public Long getAttrId();
     
     /**
      * Sets the record id.
      * @param attrId
      */
-    public void setAttrId(Integer attrId);  
+    public void setAttrId(Long attrId);  
 
     /**
      * Returns that if it is a string, otherwise null
